module_adapter: modules: Change module entry point type to uintptr_t
Change type of the entry_point parameter in the native_system_agent_start
and system_agent_start functions to uintptr_t which allows safe pointer
passing.

+++ src/audio/module_adapter/library/native_system_agent.c
@@ -19,7 +19,7 @@ typedef void* (*native_create_instance_f)(void *mod_cfg, void *parent_ppl,
 
 struct native_system_agent native_sys_agent;
 
-int native_system_agent_start(uint32_t entry_point, uint32_t module_id, uint32_t instance_id,
+int native_system_agent_start(uintptr_t entry_point, uint32_t module_id, uint32_t instance_id,
 			      uint32_t core_id, uint32_t log_handle, void *mod_cfg,
 			      const void **iface)
 {
